Move hard deletion code to BlobService.
